PUBDEV-5029: GLM crashes whem there are too many active predictors
For the multinomial case we need to consider the total of active
predictors for all classes to pick the correct solver.

If total number of active predictors for the first lambda is > 5000
and solver L_BFGS was not auto-selected we will cap the number
of predictors to 5000. This will cause the algorithm to crash for the
next lambda because the number of active predictors will violate the
invariant #active < #max.